+++ wordless/wl_utils/wl_misc.py
@@ -141,7 +141,7 @@ def wl_download_file_size(main, url):
     # In megabytes
     return file_size / 1024 / 1024
 
-def log_timing(func):
+def log_time(func):
     def wrapper(widget, *args, **kwargs):
         if isinstance(widget, QMainWindow):
             main = widget
@@ -152,20 +152,21 @@ def log_timing(func):
 
         return_val = func(widget, *args, **kwargs)
 
-        time_elapsed = time.time() - time_start
-        time_elapsed_mins = int(time_elapsed // 60)
-        time_elapsed_secs = time_elapsed % 60
+        if return_val != 'skip_logging_time':
+            time_elapsed = time.time() - time_start
+            time_elapsed_mins = int(time_elapsed // 60)
+            time_elapsed_secs = time_elapsed % 60
 
-        msg_min = _tr('wl_misc', 'minute') if time_elapsed_mins == 1 else _tr('wl_misc', 'minutes')
-        msg_time = _tr('wl_misc', '(In {} {} {:.2f} seconds)').format(time_elapsed_mins, msg_min, time_elapsed_secs)
+            msg_min = _tr('wl_misc', 'minute') if time_elapsed_mins == 1 else _tr('wl_misc', 'minutes')
+            msg_time = _tr('wl_misc', '(In {} {} {:.2f} seconds)').format(time_elapsed_mins, msg_min, time_elapsed_secs)
 
-        if (msg_cur := main.statusBar().currentMessage()):
-            if _tr('wl_misc', '(In') in msg_cur:
-                main.statusBar().showMessage(f"{msg_cur.split(_tr('wl_misc', '(In'))[0].rstrip()} {msg_time}")
+            if (msg_cur := main.statusBar().currentMessage()):
+                if _tr('wl_misc', '(In') in msg_cur:
+                    main.statusBar().showMessage(f"{msg_cur.split(_tr('wl_misc', '(In'))[0].rstrip()} {msg_time}")
+                else:
+                    main.statusBar().showMessage(f'{msg_cur} {msg_time}')
             else:
-                main.statusBar().showMessage(f'{msg_cur} {msg_time}')
-        else:
-            main.statusBar().showMessage(msg_time)
+                main.statusBar().showMessage(msg_time)
 
         return return_val