+++ wordless/wl_wordlist_generator.py
@@ -32,7 +32,7 @@ from wordless.wl_figs import wl_figs, wl_figs_freqs, wl_figs_stats
 from wordless.wl_measures import wl_measure_utils
 from wordless.wl_nlp import wl_texts, wl_token_processing
 from wordless.wl_utils import wl_conversion, wl_misc, wl_sorting, wl_threading
-from wordless.wl_widgets import wl_layouts, wl_tables, wl_widgets
+from wordless.wl_widgets import wl_boxes, wl_layouts, wl_tables, wl_widgets
 
 _tr = QCoreApplication.translate
 
@@ -44,7 +44,7 @@ class Wrapper_Wordlist_Generator(wl_layouts.Wl_Wrapper):
         self.table_wordlist_generator = Wl_Table_Wordlist_Generator(self)
 
         layout_results = wl_layouts.Wl_Layout()
-        layout_results.addWidget(self.table_wordlist_generator.label_number_results, 0, 0)
+        layout_results.addWidget(self.table_wordlist_generator.label_num_results, 0, 0)
         layout_results.addWidget(self.table_wordlist_generator.button_results_filter, 0, 2)
         layout_results.addWidget(self.table_wordlist_generator.button_results_search, 0, 3)
 
@@ -172,16 +172,17 @@ class Wrapper_Wordlist_Generator(wl_layouts.Wl_Wrapper):
 
         self.label_rank = QLabel(self.tr('Rank:'), self)
         (
+            self.checkbox_rank_sync,
             self.label_rank_min,
             self.spin_box_rank_min,
             self.checkbox_rank_min_no_limit,
             self.label_rank_max,
             self.spin_box_rank_max,
             self.checkbox_rank_max_no_limit
-        ) = wl_widgets.wl_widgets_filter(
+        ) = wl_boxes.wl_spin_boxes_min_max_no_limit(
             self,
-            filter_min = 1,
-            filter_max = 100000
+            val_min = 1,
+            val_max = 100000
         )
 
         self.combo_box_graph_type.currentTextChanged.connect(self.fig_settings_changed)
@@ -212,7 +213,8 @@ class Wrapper_Wordlist_Generator(wl_layouts.Wl_Wrapper):
 
         self.group_box_fig_settings.layout().addWidget(wl_layouts.Wl_Separator(self), 1, 0, 1, 3)
 
-        self.group_box_fig_settings.layout().addWidget(self.label_rank, 2, 0, 1, 3)
+        self.group_box_fig_settings.layout().addWidget(self.label_rank, 2, 0, 1, 2)
+        self.group_box_fig_settings.layout().addWidget(self.checkbox_rank_sync, 2, 2)
         self.group_box_fig_settings.layout().addWidget(self.label_rank_min, 3, 0)
         self.group_box_fig_settings.layout().addWidget(self.spin_box_rank_min, 3, 1)
         self.group_box_fig_settings.layout().addWidget(self.checkbox_rank_min_no_limit, 3, 2)
@@ -348,7 +350,7 @@ class Wl_Table_Wordlist_Generator(wl_tables.Wl_Table_Data_Filter_Search):
             enable_sorting = True
         )
 
-    @wl_misc.log_timing
+    @wl_misc.log_time
     def generate_table(self):
         if self.main.settings_custom['wordlist_generator']['token_settings']['assign_pos_tags']:
             nlp_support = wl_checks_work_area.check_nlp_support(self.main, nlp_utils = ['pos_taggers'])
@@ -518,7 +520,7 @@ class Wl_Table_Wordlist_Generator(wl_tables.Wl_Table_Data_Filter_Search):
             finally:
                 wl_checks_work_area.check_err_table(self.main, err_msg)
 
-    @wl_misc.log_timing
+    @wl_misc.log_time
     def generate_fig(self):
         if self.main.settings_custom['wordlist_generator']['token_settings']['assign_pos_tags']:
             nlp_support = wl_checks_work_area.check_nlp_support(self.main, nlp_utils = ['pos_taggers'])